feat: 변경된 브리지 로직을 사용하여 useBridge 로직 변경
#53
